Further refactor lossless encoder
