+++ src/lossless/encoder/bit_writer.rs
@@ -1,0 +1,46 @@
+use std::io::{self, Write};
+
+pub(crate) struct BitWriter<W> {
+    writer: W,
+    buffer: u64,
+    nbits: u8,
+}
+
+impl<W: Write> BitWriter<W> {
+    pub(crate) fn new(writer: W) -> Self {
+        Self {
+            writer,
+            buffer: 0,
+            nbits: 0,
+        }
+    }
+
+    pub(crate) fn write_bits(&mut self, bits: u64, nbits: u8) -> io::Result<()> {
+        debug_assert!(nbits <= 64);
+
+        self.buffer |= bits << self.nbits;
+        self.nbits += nbits;
+
+        if self.nbits >= 64 {
+            self.writer.write_all(&self.buffer.to_le_bytes())?;
+            self.nbits -= 64;
+            self.buffer = bits.checked_shr(u32::from(nbits - self.nbits)).unwrap_or(0);
+        }
+        debug_assert!(self.nbits < 64);
+        Ok(())
+    }
+
+    pub(crate) fn flush(&mut self) -> io::Result<()> {
+        if self.nbits % 8 != 0 {
+            self.write_bits(0, 8 - self.nbits % 8)?;
+        }
+        if self.nbits > 0 {
+            self.writer
+                .write_all(&self.buffer.to_le_bytes()[..self.nbits as usize / 8])
+                .unwrap();
+            self.buffer = 0;
+            self.nbits = 0;
+        }
+        Ok(())
+    }
+}
